backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pymupdf>=1.23.8
chromadb>=0.4.0
openai>=1.3.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pillow>=10.0.0
pytesseract>=0.3.10
aiofiles>=23.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langchain-text-splitters>=0.0.1